+++ src/www/js/index.js
@@ -1,17 +1,30 @@
-var labelVariantValue = 'Large';
+var selectedLabelVariant = 'Large';
 const tableListener = function () {
   processInput('table', 'lastRow', this);
 };
 updateTableListeners();
 
-const tableEmptyRow = `<tr><td><input type="text" class="form-control"></td><td><input type="text" class="form-control"></td><td><input type="text" class="form-control"></td><td><input type="text" class="form-control"></td></tr>`
-
+const tableEmptyRow = `
+<tr>
+  <td>
+    <input type="text" class="form-control">
+  </td>
+  <td>
+    <input type="text" class="form-control">
+  </td>
+  <td>
+    <input type="text" class="form-control">
+  </td>
+  <td>
+    <input type="text" class="form-control">
+  </td>
+</tr>`;
 
 /**
  * Run by processInput() when the inputs are updated.
  *
  * @param {string} field - The field that was updated.
- * @param {*} params - Extra params
+ * @param {*} [params] - Extra params
  */
 function updatePreview(field, ...params) {
   var value = document.getElementById(field).value;
@@ -151,7 +164,7 @@ function clearTable() {
 }
 
 /**
- * Called by the print button
+ * Called by the print buttons
  *
  * @param {boolean} whiteOnBlack - True to print white on black, false to print black on white
  * @param {boolean} [dbOnly] - True to only update the database
@@ -161,7 +174,7 @@ function print(whiteOnBlack, dbOnly) {
   let table = readTable(whiteOnBlack);
 
   if (checkTable(table)) {
-    sendForPrint(table, 'testTag', labelVariantValue, whiteOnBlack, dbOnly)
+    sendForPrint(table, 'testTag', selectedLabelVariant, whiteOnBlack, dbOnly)
   } else {
     return false;
   }