+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Program now waits for printing to finish before returning success to client
 - If barcode start and end are swapped, the program will now swap them back
+- Barcode start and end can now be 0 (or negative I guess)
 
 ## [1.1.1] - 2023-8-11